Move Go implementation to internal/
