+++ internal/interpreter/parser/util.go
