+++ internal/jsone_test.go
@@ -56,7 +56,7 @@ func TestSpec(t *testing.T) {
 	// in gopkg.in/yaml.v2
 
 	// Read specification.yml
-	data, err := ioutil.ReadFile("specification.yml")
+	data, err := ioutil.ReadFile("../specification.yml")
 	require.NoError(t, err, "failed to read specification.yml")
 	// Parse as YAML (split for each document)
 	var rawSpec []interface{}
